+++ tech1-framework-incidents/src/main/java/io/tech1/framework/incidents/Abc.java
@@ -1,4 +1,0 @@
-package io.tech1.framework.incidents;
-
-public interface Abc {
-}