+++ tech1-framework-incidents/src/main/java/io/tech1/framework/incidents/domain/Incident.java
@@ -1,0 +1,69 @@
+package io.tech1.framework.incidents.domain;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import io.tech1.framework.domain.base.Username;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.ToString;
+
+import java.util.Map;
+import java.util.TreeMap;
+
+import static io.tech1.framework.domain.asserts.Asserts.assertNonNullOrThrow;
+import static io.tech1.framework.domain.constants.StringConstants.UNKNOWN;
+import static io.tech1.framework.domain.utilities.exceptions.ExceptionsMessagesUtility.invalidAttribute;
+import static java.util.Objects.isNull;
+import static java.util.Objects.nonNull;
+
+// Lombok
+@Getter
+@EqualsAndHashCode
+@ToString
+public class Incident {
+    private Map<String, Object> attributes;
+
+    public Incident() {
+        this.attributes = new TreeMap<>();
+    }
+
+    public static Incident copyOf(Incident incident) {
+        assertNonNullOrThrow(incident, invalidAttribute("Incident.incident"));
+        var instance = new Incident();
+        incident.getAttributes().forEach(instance::add);
+        return instance;
+    }
+
+    public void add(String key, Object value) {
+        if (isNull(this.attributes)) {
+            this.attributes = new TreeMap<>();
+        }
+        this.attributes.put(key, value);
+    }
+
+    public void addAll(Map<String, Object> attributes) {
+        if (isNull(this.attributes)) {
+            this.attributes = new TreeMap<>();
+        }
+        this.attributes.putAll(attributes);
+    }
+
+    @JsonIgnore
+    public String getType() {
+        var attribute = this.attributes.get(IncidentAttributes.Keys.TYPE);
+        if (nonNull(attribute)) {
+            return attribute.toString();
+        } else {
+            return UNKNOWN;
+        }
+    }
+
+    @JsonIgnore
+    public Username getUsername() {
+        var attribute = this.attributes.get(IncidentAttributes.Keys.USERNAME);
+        if (nonNull(attribute)) {
+            return Username.of(attribute.toString());
+        } else {
+            return Username.of(UNKNOWN);
+        }
+    }
+}