+++ tech1-framework-incidents/src/main/java/io/tech1/framework/incidents/domain/authetication/AuthenticationLoginFailureUsernamePasswordIncident.java
@@ -1,0 +1,11 @@
+package io.tech1.framework.incidents.domain.authetication;
+
+import io.tech1.framework.domain.base.Username;
+import lombok.Data;
+
+// Lombok
+@Data(staticConstructor = "of")
+public class AuthenticationLoginFailureUsernamePasswordIncident {
+    private final Username username;
+    private final String password;
+}