+++ tech1-framework-incidents/src/main/java/io/tech1/framework/incidents/domain/registration/Register1Incident.java
@@ -1,0 +1,10 @@
+package io.tech1.framework.incidents.domain.registration;
+
+import io.tech1.framework.domain.base.Username;
+import lombok.Data;
+
+// Lombok
+@Data(staticConstructor = "of")
+public class Register1Incident {
+    private final Username username;
+}