+++ tech1-framework-incidents/src/main/java/io/tech1/framework/incidents/feigns/definitions/IncidentClientDefinition.java
@@ -1,0 +1,12 @@
+package io.tech1.framework.incidents.feigns.definitions;
+
+import feign.Headers;
+import feign.RequestLine;
+import io.tech1.framework.incidents.domain.Incident;
+import org.springframework.http.MediaType;
+
+public interface IncidentClientDefinition {
+    @RequestLine("POST /api/incidents/register")
+    @Headers("Content-Type: " + MediaType.APPLICATION_JSON_VALUE)
+    void registerIncident(Incident incident);
+}